+++ examples/tabular/example_custom_model_tabular.py
@@ -1,10 +1,10 @@
 """ Example script for defining and using custom models in AutoGluon Tabular """
 
+from autogluon.core.utils import infer_problem_type
 from autogluon.tabular import TabularDataset, TabularPredictor
 from autogluon.tabular.configs.hyperparameter_configs import get_hyperparameter_config
 from autogluon.tabular.data.label_cleaner import LabelCleaner
 from autogluon.tabular.models import AbstractModel
-from autogluon.tabular.utils import infer_problem_type
 
 
 #########################
