+++ tabular/src/autogluon/tabular/trainer/abstract_trainer.py
@@ -15,8 +15,8 @@ from autogluon.core.utils.exceptions import TimeLimitExceeded, NotEnoughMemoryEr
 from autogluon.core.utils import default_holdout_frac
 from autogluon.core.metrics import scorer_expects_y_pred
 from autogluon.core.scheduler.scheduler_factory import scheduler_factory
+from autogluon.core.utils import get_pred_from_proba, generate_train_test_split, infer_eval_metric, compute_permutation_feature_importance
 
-from ..utils import get_pred_from_proba, generate_train_test_split, infer_eval_metric, compute_permutation_feature_importance
 from ..models.abstract.abstract_model import AbstractModel
 from ..models.ensemble.bagged_ensemble_model import BaggedEnsembleModel
 from ..trainer.model_presets.presets_custom import get_preset_custom
